+++ source/world/world.cpp
@@ -31,7 +31,7 @@ namespace terminus
 World::World(Game & game,
              bool isNetworkGame,
              bool isPlayerOne,
-             unsigned int terrainSeed,
+             long long terrainSeed,
              QList<WagonType> playerWagons,
              QList<WagonType> enemyWagons)
 : m_game(game)
